Interprocess Communication using Pipes
A full duplex communication between parent and child process.
Parent process writes Filename of a file on pipe1.
This filename is read by child process.
The child process writes the contents of the file to pipe 2.
The contents are read from pipe2 by the parent process and is displayed on standard output.